chore(clippy): clear the workspace lint debt unmasked by the fp-columnar fix
`cargo clippy --workspace --all-targets -- -D warnings` had been failing
at fp-columnar, which masked ~260 accumulated errors in fp-frame/fp-io
behind it (CI's lint gate never reached those crates). With fp-columnar
clean, this pass takes the whole workspace to a green lint gate:

- fp-frame: remove 215 no-op `.into()` conversions in DataFrame struct
  literals (`columns` / `column_order` already have the target types)
  and the redundant field names that exposed; `iter().any(|&v| v == X)`
  -> `contains(&X)` (11 sites); collapse the issue-#19 ewm fast-path
  gates and three groupby else-if arms into let-chains; `sort_by` ->
  `sort_by_key` (3 sites, stability preserved); needless range loops ->
  iterator forms; `&ls[..]` -> `ls`; `s[a..b].as_bytes()` ->
  `&s.as_bytes()[a..b]`; escape a doc `>` false quote; manual clamp ->
  `clamp`; constant assert -> `const` block; digit-grouping and
  parenthesization nits in examples; `#[allow(clippy::type_complexity)]`
  on `as_mapping`.
- fp-frame: replace ALL `vec![Vec::with_capacity(n); k]` (8 sites) with
  per-element construction — the macro clones, so every clone silently
  had capacity 0 and the intended preallocation never happened
  (clippy repeat_vec_with_capacity; latent perf bug, no behavior change).
- fp-frame tests: sqrt(2)/1/sqrt(2) literals -> `std::f64::consts::
  {SQRT_2, FRAC_1_SQRT_2}` (bit-identical values).
- fp-io: doc list continuation fix, `#[allow(clippy::type_complexity)]`
  on a scoped-thread result binding.
- examples across fp-frame/fp-io/fp-index/fp-join: machine-applicable
  `cargo clippy --fix` output (needless_range_loop, uninlined format
  args, etc.) plus `cargo fmt`.

Gates at this commit: `cargo clippy --workspace --all-targets --
-D warnings` exit 0; `cargo fmt --check` exit 0.

+++ crates/fp-frame/examples/bench_dum_dig.rs
@@ -13,7 +13,7 @@ fn main() {
     let mut d = 0u64;
     for seed in 0..40u64 {
         let n = (sm(seed as usize, 3) % 2000 + 1) as usize;
-        let card = (sm(seed as usize, 4) % 20 + 1) as u64;
+        let card = sm(seed as usize, 4) % 20 + 1;
         let mut m = BTreeMap::new();
         m.insert(
             "c".to_string(),
